# hdkeychain: Use secp256k1 privkey to pubkey method.